Added Scripts menu
Added Scripts menu in the main System Settings page. The menu scans for all *.sh files in the SD and, if one is selected, it executes it showing the standard output in the user interface; a Cancel option allows to interrupt the script and becomes OK once the script has finished.