makefile: Make obj and dep rules unique per language
The make rules for objects and dependency files must be unique per
language or make won't understand when to rebuild them.  Before this
change, if a cpp-file was changed, make would match it to the rule for C
files because it comes first in the `Makefile` and not rebuild anything.