Implemented type-to-filter functionality in long game lists
Previously, when you started typing the name of a game in the long list of
games, it would jump to each successive letter you typed. So if you wanted to
play "Crash Bandicoot" on the GBA, and you started typing "CRA" on the keyboard,
the menu would first jump to C, then R, then A. This is almost certainly not
what you want, and not standard UX for these kinds of long lists.

This commit changes the behaviour to filter the list based on what you've typed.
So if you type "CRA", it will only show games beginning with "CRA". If you wish
to clear the filter, you can do so with backspace. Also: if you don't type for a
little bit (currently 2000 ms), it will also clear the prefix the next time you
start typing.

The functionality is implemented using the "prefix" argument to ScanDirectory.