Use proper filter instead of prefix to filter file lists
Added a new argument to ScanDirectory, a "filter" parameter. If it exists, it
will filter items regardless of where in the filename the filter appears (not
just in the beginning, as with the prefix). So if you type "ADV" as the filter,
both "Advance Wars" as well as "Ace Combat Advance" will pass through