minimig: Fix path resolution for assigns
Only nullify the key in find_path() if the name contains our device
(SHARE:) or volume (MiSTer:) name.

This fixes the ability to create assigns to locations on the shared
filesystem. If find_path() is called with an assign-relative path, we
want to preserve the key so that a relative path is constructed from the
directory associated with that key.