Change sorting comparison to generate expected results (#466)
Currently when you have two entries where one has a longer name ("foobar" and "foobar 2") it does the comparison up to the shorter of the lengths, and then if they match to that length, it does a date comparison instead. This exhibits behaviour different to what one would expect from most implementations.

This change puts the length comparison before the date comparison to make it more like the user would expect.

i.e "foobar" should always be above "foobar 2", and not move depending on the date code.