lastgame bootcore feature
# Description
This change expands the `bootcore` ini option to support a new mode: "lastgame". The lastgame mode will load the previously loaded core (just like "lastcore") but it will also attempt to load the ROM that was previously being played on that core also.

## user_io and menu.

The work of parsing the confstr for file and mount descriptions and handling the loading/mount of them has been moved from the menu code to user_io.

`user_io_parse_selection` parses a confstr entry and fills a `uio_selection_descriptor` with the necessary information for the file or image. The parsing code in the menu system has been replaced by this and several of the globals and function statics (`opensave`, `load_addr`, `ioctl_index`, `store_name`) have been replaced by a static `sel_desc` instance of a `uio_selection_descriptor` which now contains the same information. The `user_io_load_or_mount` is new functionality, it works in reverse, given a filepath it determines which confstr entry matches it's extension and loads or mounts it appropriately.

The code in the `MENU_GENERIC_FILE_SELECTED` and `MENU_GENERIC_IMAGE_SELECTED` that was responsible for loading/mounting the selected file(s) has been moved to `user_io_load_file` and `user_io_mount_image` respectively. The menu code is still responsible for handling the "UI" specfic tasks like recording recent files.

These changes to the menu implementation this need a lot more testing. For instance, the archie and minimig specific code were setting `ioctl_index` but it seems like this was not actually being used for anything and mounting multiple disk images on both of these cores works without issues.

## bootcore
Rewrote most of this because it needs now needs to handle more state and I didn't want to work around how some of it previous worked. Internal functions are no longer in the header and the bootcore state is handled internally (moved from menu.cpp) and doesn't modify the cfg data anymore. The `lastcore.dat` data that gets written out is not a structure that contains the core path along with the name of the core and the last file that was loaded for the core. When a file is selected and when store_name is false (since those files would be loaded automatically when the core starts anyway) that path is written to the lastcore data via `bootcore_record_file`. On core startup the `bootcore_load_file` function will attempt to load that file via the `user_io_load_or_mount` function. Currently only files (ROMS) are tracked, the lastgame functionality doesn't make any attempt to mount disk images.

## Misc
Changes to support/* and file_io.* are just function signature fixes required for const correctness.