Change "Analog" to "Stick" to match joy_ana_map
The current strings for `joy_button_map` are not as intuitive to the user as they probably could be. This proposed change will help the user know what is meant by the defining the axis of the joystick by referring to the stick the same way as it was referred to during the earlier controller test step. If there is a better way to do this or this is not appropriate, that's okay too. :)