Keyboard scancode set switching and set 1 support
This improvement allows the use of set 1 of keyboard scancodes, as well as the possibility of changing from one to another from the OS. Both core ao486 and new cores based on set 1 will benefit from this feature.

Thank you @naeloob for your cooperation and support in developing this feature.