Account for aspect ratio in vscale_mode 4/5 (#620)
Both modes ensure there is enough horiztonal resolution to fit the core.
They do this by reducing the vertical resolution as needed. Mode 5 keeps
the horizontal resolution fixed, which uses more bandwidth but reduces
the amount of resolution switches that can occur and prevents the OSD
from getting to narrow in rotated arcade cores.

Initialize the video_info with the current data if the fb_crc changes to
ensure it has valid data for the resolution if the resolution has not
changed.